+++ eth_defi/uniswap_v3/analysis.py
@@ -1,3 +1,5 @@
+from typing import TypedDict
+
 from web3 import Web3
 from web3.logs import DISCARD
 
@@ -10,7 +12,7 @@ from eth_defi.uniswap_v3.pool import fetch_pool_details
 from eth_defi.uniswap_v3.utils import decode_path
 
 
-def get_input_args(params: tuple | dict) -> dict:
+def get_input_args(params: tuple | dict) -> tuple:
     """Names and decodes input arguments from router.decode_function_input()
     Note there is no support yet for SwapRouter02, it does not accept a deadline parameter
     See: https://docs.uniswap.org/contracts/v3/reference/periphery/interfaces/ISwapRouter#exactinputparams
@@ -19,7 +21,7 @@ def get_input_args(params: tuple | dict) -> dict:
     params from router.decode_function_input
 
     :returns:
-    Dict of exactInputParams as specified in the link above
+        Dict of exactInputParams as specified in the link above
     """
 
     if type(params) == dict:
@@ -50,7 +52,26 @@ def analyse_trade_by_receipt(
     tx: dict,
     tx_hash: str | bytes,
     tx_receipt: dict,
+    input_args: tuple | None = None,
 ) -> TradeSuccess | TradeFail:
+    """Analyse a Uniswpa v3 trade.
+
+    Figure out
+
+    - The success of the trade
+
+    - Slippage, etc.
+
+
+    :param tx_receipt:
+        Transaction receipt
+
+    :param input_args:
+        The swap input arguments.
+
+        If not given automatically decode from `tx`.
+        You need to pass this for Enzyme transactions, because transaction payload is too complex to decode.
+    """
     router = uniswap.swap_router
     assert tx_receipt["to"] == router.address, f"For now, we can only analyze naive trades to the router. This tx was to {tx_receipt['to']}, router is {router.address}"
 
@@ -63,14 +84,20 @@ def analyse_trade_by_receipt(
         reason = fetch_transaction_revert_reason(web3, tx_hash)
         return TradeFail(gas_used, effective_gas_price, revert_reason=reason)
 
-    # Decode inputs going to the Uniswap swap
-    # https://stackoverflow.com/a/70737448/315168
-    function, params_struct = router.decode_function_input(get_transaction_data_field(tx))
-    input_args = get_input_args(params_struct["params"])
+    if input_args is None:
+        # Decode inputs going to the Uniswap swap
+        # https://stackoverflow.com/a/70737448/315168
+        function, params_struct = router.decode_function_input(get_transaction_data_field(tx))
+        input_args = get_input_args(params_struct["params"])
+        assert function.fn_name == "exactInput", f"Unsupported Uniswap v3 trade function {function}"
+    else:
+        # Decode from Enzyme stored input
+        # Note that this is how Web3.py presents this
+        # <Function exactInput((bytes,address,uint256,uint256,uint256)) bound to ((b"'\x91\xbc\xa1\xf2\xdeFa\xed\x88\xa3\x0c\x99\xa7\xa9D\x9a\xa8At\x00\x01\xf4\rP\x0b\x1d\x8e\x8e\xf3\x1e!\xc9\x9d\x1d\xb9\xa6DM:\xdf\x12p", '0xfC3035f60A3d862E0753eA3D2Eec7679227E8B37', 9223372036854775808, 1000000, 1144586690647966336),)>
+        input_args = get_input_args(input_args[0])
 
     path = input_args["path"]
 
-    assert function.fn_name == "exactInput", f"Unsupported Uniswap v3 trade function {function}"
     assert len(path), f"Seeing a bad path Uniswap routing {path}"
 
     amount_in = input_args["amountIn"]