+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add [Aave v3 deployer support](https://github.com/aave/aave-v3-deploy) in`eth_defi.aave_v3.deployer` module
 - Add Solidity library linking support for Hardhat-based deployments in `eth_defi.abi.get_linked_contract`
 - Add: More logging and better error messages to some parts 
+- Add: `figure_reorganisation_and_new_blocks(max_range)` to catch situations you somehow feed a too long block range to scan
+- Add: `analyse_trade_by_receipt(input_args)` to analyse the success of Uni v3 trades when trading on Enzyme
 
 # 0.18.4